int to float

+++ api/controllers/SensorsController.js
@@ -9,7 +9,7 @@ module.exports = {
 	changeTemp: function(req,res) {
 		Sensors.findOne({name: req.query.name}).exec(function(err,ret){
 			if(err) return res.send(err);
-			Sensors.update({name: req.query.name}, {temp:parseInt(req.query.temp)}).exec(function(err,ret){
+			Sensors.update({name: req.query.name}, {temp:parseFloat(req.query.temp)}).exec(function(err,ret){
 				if(err) return res.send(err);
 				return res.send(ret);
 			});
@@ -18,7 +18,7 @@ module.exports = {
 	changeLight: function(req, res) {
 		Sensors.findOne({name: req.query.name}).exec(function(err,ret){
 			if(err) return res.send(err);
-			Sensors.update({name: req.query.name}, {temp:parseInt(req.query.light)}).exec(funtion(err,ret){
+			Sensors.update({name: req.query.name}, {temp:parseFloat(req.query.light)}).exec(funtion(err,ret){
 				if(err) return res.send(err);
 				return res.send(ret);
 			});
@@ -27,7 +27,7 @@ module.exports = {
 	changeTime: function(req,res) {
 		Sensors.findOne({name: "global").exec(function(err,ret){
 			if(err) return res.send(err);
-			Sensors.update({name: "global"}, {time:parseInt(req.query.time)}).exec(funtion(err,ret){
+			Sensors.update({name: "global"}, {time:parseFloat(req.query.time)}).exec(funtion(err,ret){
 				if(err) return res.send(err);
 				return res.send(ret);
 			});
